requirements.txt
openai==1.12.0
PyPDF2==3.0.1
selenium==4.31.0
webdriver-manager==4.0.2
python-pptx==1.0.2
beautifulsoup4==4.13.4
pyttsx3==2.98
ffmpeg-python==0.2.0
regex==2024.11.6
requests==2.32.3
numpy==2.0.2
Pillow==10.4.0
lxml==5.3.2
tqdm==4.67.1
python-dotenv==1.1.0
urllib3==2.4.0
protobuf==5.29.4
pydantic==2.11.3
typing_extensions==4.13.2
pywin32==310
pypiwin32==223
comtypes==1.4.10
moviepy==1.0.3
opencv-python==4.11.0.86
google-generativeai==0.8.5
azure-cognitiveservices-speech==1.43.0